changes in README, slightly refactored dag functions. Simplify parts of the installation. Still it's not working for me.
