+++ api_functions.py
@@ -6,7 +6,6 @@ file to be copied in airflow/dags
 
 
 import requests
-from pulp import *
 
 def sign_up(email, pwd, name):
     return requests.post(
@@ -48,23 +47,6 @@ def get_data(token, execution_id):
     return response.json()
 
 
-def solve_model(data, config):
-    print("Solving the model")
-    var, model = LpProblem.from_dict(data)
-    print(config)
-    solver = get_solver_from_dict(config)
-    model.solve(solver)
-    solution = model.to_dict()
-    
-    log_path = config["logPath"]
-    f = open(log_path, "r")
-    log = f.read()
-    
-    print("Model solved")
-    
-    return solution, log
-
-
 def write_solution(token, execution_id, solution, log_text=None, log_json=None):
     print("Writing the solution in database")
     response = requests.post(
@@ -75,12 +57,7 @@ def write_solution(token, execution_id, solution, log_text=None, log_json=None):
     return response
 
 
-def solve_execution(token, execution_id):
-    execution_data = get_data(token, execution_id)
-    solution, log = solve_model(execution_data["data"], execution_data["config"])
-    write_solution(token, execution_id, solution, log)
-    
-    return solution
+
