+++ requirements.txt
@@ -14,7 +14,7 @@ itsdangerous==1.1.0
 Jinja2==2.11.2
 Mako==1.1.2
 MarkupSafe==1.1.1
-marshmallow==3.6.0
+marshmallow==2.18.0
 psycopg2==2.8.5
 pycparser==2.20
 PyJWT==1.7.1
@@ -24,3 +24,4 @@ pytz==2020.1
 six==1.14.0
 SQLAlchemy==1.3.17
 Werkzeug==1.0.1
+pulp==2.2
